refactor: 사용되지 않는 커스텀 예외 클래스 제거 및 패키지 정리
