--- backend/src/main/java/com/woowacourse/kkogkkog/exception/auth/AccessTokenRetrievalFailedException.java
@@ -1,10 +1,0 @@
-package com.woowacourse.kkogkkog.exception.auth;
-
-import com.woowacourse.kkogkkog.exception.UnauthenticatedException;
-
-public class AccessTokenRetrievalFailedException extends UnauthenticatedException {
-
-    public AccessTokenRetrievalFailedException(String message) {
-        super(message);
-    }
-}
