+++ backend/src/test/java/com/woowacourse/kkogkkog/auth/application/AuthServiceTest.java
@@ -10,7 +10,7 @@ import static org.mockito.BDDMockito.given;
 import com.woowacourse.kkogkkog.support.setup.ServiceTest;
 import com.woowacourse.kkogkkog.member.application.dto.MemberResponse;
 import com.woowacourse.kkogkkog.auth.application.dto.TokenResponse;
-import com.woowacourse.kkogkkog.exception.auth.AccessTokenRetrievalFailedException;
+import com.woowacourse.kkogkkog.exception.infrastructure.AccessTokenRetrievalFailedException;
 import com.woowacourse.kkogkkog.infrastructure.dto.SlackUserInfo;
 import com.woowacourse.kkogkkog.infrastructure.dto.WorkspaceResponse;
 import org.junit.jupiter.api.DisplayName;
@@ -59,7 +59,7 @@ class AuthServiceTest extends ServiceTest {
         @DisplayName("올바르지 않은 임시 코드를 입력하면, 예외를 던진다")
         void fail_invalidCode() {
             given(slackClient.getUserInfoByCode("invalid_code"))
-                .willThrow(new AccessTokenRetrievalFailedException("invalid_code"));
+                .willThrow(new AccessTokenRetrievalFailedException());
 
             assertThatThrownBy(
                 () -> authService.login("invalid_code")
